refactor: Add request preloads, minor style changes
Playing with Network Waterfall, parallel requests to improve page loading times.
The 'header-loader' is now fixed at the top of the header instead of its bottom.
CSRF '<input>' tag is now changed to '<meta>' tag.
The OpenCode's logo is now reduced in size as it was redundant to use a high-res image.